% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{DIS_Integrate_Profile}
\alias{DIS_Integrate_Profile}
\title{2nd function for cholorphyll integration}
\usage{
DIS_Integrate_Profile(DEPH, value, depth.max, depth_range)
}
\arguments{
\item{DEPH}{helper function}

\item{value}{helper function}

\item{depth.max}{helper function}

\item{depth_range}{helper function}
}
\value{
helper functions
}
\description{
2nd function for cholorphyll integration
}
