% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{DIS_Profile_Integration}
\alias{DIS_Profile_Integration}
\title{1st function for cholorphyll integration}
\usage{
DIS_Profile_Integration(df.data, depth_range)
}
\arguments{
\item{df.data}{bottle data}

\item{depth_range}{vector of 2 values}
}
\value{
helper function
}
\description{
1st function for cholorphyll integration
}
