% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{set_sgdo_pass}
\alias{set_sgdo_pass}
\title{SGDO credentials}
\usage{
set_sgdo_pass()
}
\value{
your credential are stored into the otions. Can be use with this function or by loading a Rprofile file that set options.This allows for script sharing without sharing credentials.Options should include SGDO_host, SGDO_uid, SGDO_pass
}
\description{
SGDO credentials
}
