% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{SGDO_sql}
\alias{SGDO_sql}
\title{SQL query and summary of environmental variables using SGDO}
\usage{
SGDO_sql(
  drv,
  latitude,
  longitude,
  date,
  time,
  timezone = "UTC",
  depth.max = Inf,
  ID,
  geotol = 0.02,
  timetol = 24,
  station = NULL,
  range_ctd = c(0, 50),
  range_bot = c(0, 100)
)
}
\arguments{
\item{drv}{driver use this : DBI::dbDriver("Oracle")}

\item{latitude}{vector of latitude please remove na from data frame prior to function}

\item{longitude}{vector of longitude, please remove na from data frame prior to function}

\item{date}{vector of date format year month day}

\item{time}{vector of time format HH:MM}

\item{timezone}{vector of time zone, please use time zone recongnizable by R}

\item{depth.max}{vector of depth of station. for now it is requisited but may become optional in the future}

\item{ID}{unique identifier of stations/rows.}

\item{geotol}{geographic tolerance for searching profiles near the station. Default to 0.1 degrees.}

\item{timetol}{time tolerance for searching profiles near the station, default to 24}

\item{station}{consecutive identifier if available, facultative. If given, the CTD in the search radius that has the same consecutive as the station is selected.}

\item{range_ctd}{vector of 2 values, identifying the minimum and maximum range for summarizing temperature and salinity}

\item{range_bot}{vector of 2 values, for chlrophyll integration}
}
\value{
a data frame  with the unique idnetifier and calculated variables, join with your data using the unique identifier
}
\description{
SQL query and summary of environmental variables using SGDO
}
\examples{
#df <-  data.frame(latitude=47.17133,	longitude=-63.7465, time=20:33, timezone=America/Halifax, depth.max=56, ID=1, consecutive=72)
#SGDO_sql(latitude=df$latitude, longitude=df$longitude, date=df$date, time=df$time, timezone=df$timezone,depth.max=df$depth.max, ID=df$ID, geotol=0.01, timetol=6, station=df$consecutive)



}
