% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{summarize_CTD}
\alias{summarize_CTD}
\title{Summarizing CTD profiles}
\usage{
summarize_CTD(df_data, depth_range = c(0, 50))
}
\arguments{
\item{df_data}{data frame with following variables: DEPH, TE90, PSAL, DENS, depth.max}

\item{depth_range}{vector of 2 values identifying the range of temperature and salinity for average}
}
\value{
a data frame of 1 row for all variables calculated for the unique identifier
}
\description{
Summarizing CTD profiles
}
