% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Trapz}
\alias{Trapz}
\title{3rd function for cholorphyll integration}
\usage{
Trapz(x, y)
}
\arguments{
\item{x}{helper function}

\item{y}{helper function}
}
\value{
helper function
}
\description{
3rd function for cholorphyll integration
}
