% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooenv-package.R
\docType{package}
\name{zooenv-package}
\alias{zooenv}
\alias{zooenv-package}
\title{zooenv: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
