% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooenv-package.R
\docType{package}
\name{zooenv-package}
\alias{zooenv}
\alias{zooenv-package}
\title{zooenv: Environmental variables merging with zooplancton data}
\description{
Extract and compute indices from CTD and Bottle data and merge with zooplankton dataframe.
}
\keyword{internal}
